requirements.txt
langchain
langchain_community
pypdf
langchain_text_splitters
python-dotenv
langchain-openai
langchain-weaviate
weaviate-client
pymupdf 
pdfplumber
#spacy==3.8.2                     # wheels exist for CPython 3.11
# model wheel pulled directly from spaCy’s GitHub release page
#en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl